LearningSupportTeamCategory: fix error on delete item with Learning support team related

+++ app/src/Backoffice/LearningSupportTeamCategory/Application/Delete/LearningSupportTeamCategoryDeleter.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Backoffice\LearningSupportTeamCategory\Application\Delete;
 
 use App\Backoffice\LearningSupportTeamCategory\Application\Get\Single\LearningSupportTeamCategoryFinder;
+use App\Backoffice\LearningSupportTeamCategory\Domain\Exception\CannotDeleteCategoryWithRelatedLearningSupportTeams;
 use App\Backoffice\LearningSupportTeamCategory\Domain\LearningSupportTeamCategoryRepository;
 
 final class LearningSupportTeamCategoryDeleter
@@ -22,6 +23,10 @@ final class LearningSupportTeamCategoryDeleter
     public function __invoke(string $id)
     {
         $learningSupportTeamCategory = $this->finder->__invoke($id);
+
+        if ($learningSupportTeamCategory->hasLearningSupportTeams()) {
+            throw new CannotDeleteCategoryWithRelatedLearningSupportTeams();
+        }
 
         $this->repository->delete($learningSupportTeamCategory);
     }
